lisa: Bring back dolby secure decoders and fix crash
Seems like new changes to c2 libs in android 12,
broke these decoders. Suspicious changes:
https://github.com/AOSPA/android_frameworks_av/commit/34821f568bc39602752ce056e7c4df57d12f6459
https://github.com/AOSPA/android_frameworks_av/commit/34821f568bc39602752ce056e7c4df57d12f6459
https://github.com/AOSPA/android_frameworks_av/commit/1d5b3f9539ccfd0b77a07a446cb0595c221497ce

Copy the old libs and patch dolby binary to load
it, in order to fix/workaround this problem.

While we're at it, drop separate c2 manifest and
add the dolby instance in the main manifest itself.

This reverts commits b6bf2345d3b687701ad626d34f58cf1ea4c9f966 and b9848af2ca6a87fbce95cee101d193e453e21923.

Test: Dolby Vision playback in Netflix.
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: klozz <carlosj@klozz.dev>